Add inspection visitor handler
