+++ src/main/kotlin/org/domaframework/doma/intellij/inspection/sql/visitor/SqlInspectionVisitor.kt
@@ -19,29 +19,14 @@ import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiLiteralExpression
-import com.intellij.psi.PsiMethod
-import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.elementType
-import org.domaframework.doma.intellij.common.dao.findDaoMethod
 import org.domaframework.doma.intellij.common.isInjectionSqlFile
 import org.domaframework.doma.intellij.common.isJavaOrKotlinFileType
-import org.domaframework.doma.intellij.common.psi.PsiDaoMethod
-import org.domaframework.doma.intellij.common.psi.PsiParentClass
-import org.domaframework.doma.intellij.common.psi.PsiStaticElement
-import org.domaframework.doma.intellij.common.sql.cleanString
-import org.domaframework.doma.intellij.common.sql.validator.result.ValidationClassPathResult
-import org.domaframework.doma.intellij.common.sql.validator.result.ValidationDaoParamResult
-import org.domaframework.doma.intellij.common.sql.validator.result.ValidationForDirectiveItemTypeResult
-import org.domaframework.doma.intellij.common.sql.validator.result.ValidationPropertyResult
-import org.domaframework.doma.intellij.common.util.ForDirectiveUtil
-import org.domaframework.doma.intellij.extension.expr.accessElements
-import org.domaframework.doma.intellij.extension.psi.findParameter
-import org.domaframework.doma.intellij.extension.psi.getForItem
 import org.domaframework.doma.intellij.extension.psi.isFirstElement
-import org.domaframework.doma.intellij.extension.psi.psiClassType
+import org.domaframework.doma.intellij.inspection.sql.handler.InspectionFieldAccessVisitorProcessor
+import org.domaframework.doma.intellij.inspection.sql.handler.InspectionPrimaryVisitorProcessor
+import org.domaframework.doma.intellij.inspection.sql.handler.InspectionStaticFieldAccessVisitorProcessor
 import org.domaframework.doma.intellij.psi.SqlElFieldAccessExpr
-import org.domaframework.doma.intellij.psi.SqlElForDirective
-import org.domaframework.doma.intellij.psi.SqlElIdExpr
 import org.domaframework.doma.intellij.psi.SqlElPrimaryExpr
 import org.domaframework.doma.intellij.psi.SqlElStaticFieldAccessExpr
 import org.domaframework.doma.intellij.psi.SqlTypes
@@ -65,7 +50,8 @@ class SqlInspectionVisitor(
 
     override fun visitElStaticFieldAccessExpr(element: SqlElStaticFieldAccessExpr) {
         super.visitElStaticFieldAccessExpr(element)
-        checkStaticFieldAndMethodAccess(element, holder)
+        val handler = InspectionStaticFieldAccessVisitorProcessor(this.shortName)
+        handler.check(element, holder)
     }
 
     override fun visitElFieldAccessExpr(element: SqlElFieldAccessExpr) {
@@ -73,14 +59,8 @@ class SqlInspectionVisitor(
         if (setFile(element)) return
         val visitFile: PsiFile = file ?: return
 
-        // Get element inside block comment
-        val blockElement = getFieldAccessBlocks(element)
-        val topElm = blockElement.firstOrNull() as SqlElPrimaryExpr
-
-        // Exclude fixed Literal
-        if (isLiteralOrStatic(topElm)) return
-
-        checkAccessFieldAndMethod(holder, blockElement, visitFile)
+        val handler = InspectionFieldAccessVisitorProcessor(shortName, element)
+        handler.check(holder, visitFile)
     }
 
     override fun visitElPrimaryExpr(element: SqlElPrimaryExpr) {
@@ -89,150 +69,7 @@ class SqlInspectionVisitor(
         if (setFile(element)) return
         val visitFile: PsiFile = file ?: return
 
-        if (isLiteralOrStatic(element)) return
-        PsiTreeUtil.getParentOfType(element, SqlElStaticFieldAccessExpr::class.java)?.let { return }
-
-        val forDirectiveExp = PsiTreeUtil.getParentOfType(element, SqlElForDirective::class.java)
-        val isSkip = forDirectiveExp != null && forDirectiveExp.getForItem() != element
-
-        var forDirectiveBlocks = ForDirectiveUtil.getForDirectiveBlocks(element, skipSelf = isSkip)
-        val forItem = ForDirectiveUtil.findForItem(element, skipSelf = isSkip, forDirectives = forDirectiveBlocks)
-        if (forItem != null) {
-            val forDeclarationType =
-                ForDirectiveUtil.getForDirectiveItemClassType(
-                    element.project,
-                    forDirectiveBlocks,
-                    forItem,
-                )
-            if (forDeclarationType == null) {
-                ValidationForDirectiveItemTypeResult(
-                    element,
-                    this.shortName,
-                ).highlightElement(holder)
-            }
-            return
-        }
-
-        val daoMethod = findDaoMethod(visitFile) ?: return
-        val param = daoMethod.findParameter(cleanString(element.text))
-        if (param != null) return
-
-        ValidationDaoParamResult(
-            element,
-            daoMethod.name,
-            this.shortName,
-        ).highlightElement(holder)
-    }
-
-    private fun getFieldAccessBlocks(element: SqlElFieldAccessExpr): List<SqlElIdExpr> {
-        val blockElements = element.accessElements
-        (blockElements.firstOrNull() as? SqlElPrimaryExpr)
-            ?.let { if (isLiteralOrStatic(it)) return emptyList() }
-            ?: return emptyList()
-
-        return blockElements.mapNotNull { it as SqlElIdExpr }
-    }
-
-    private fun checkAccessFieldAndMethod(
-        holder: ProblemsHolder,
-        blockElement: List<SqlElIdExpr>,
-        file: PsiFile,
-    ) {
-        val topElement = blockElement.firstOrNull() ?: return
-        val daoMethod = findDaoMethod(file) ?: return
-        val project = topElement.project
-        val forDirectiveBlocks = ForDirectiveUtil.getForDirectiveBlocks(topElement)
-        val forItem = ForDirectiveUtil.findForItem(topElement, forDirectives = forDirectiveBlocks)
-        var isBatchAnnotation = false
-        val topElementParentClass =
-            if (forItem != null) {
-                val result = ForDirectiveUtil.getForDirectiveItemClassType(project, forDirectiveBlocks, forItem)
-                if (result == null) {
-                    ValidationForDirectiveItemTypeResult(
-                        topElement,
-                        this.shortName,
-                    ).highlightElement(holder)
-                    return
-                }
-                val specifiedClassType =
-                    ForDirectiveUtil.resolveForDirectiveItemClassTypeBySuffixElement(topElement.text)
-                if (specifiedClassType != null) {
-                    PsiParentClass(specifiedClassType)
-                } else {
-                    result
-                }
-            } else {
-                val paramType = daoMethod.findParameter(cleanString(topElement.text))?.type
-                if (paramType == null) {
-                    errorHighlight(topElement, daoMethod, holder)
-                    return
-                }
-                isBatchAnnotation = PsiDaoMethod(project, daoMethod).daoType.isBatchAnnotation()
-                PsiParentClass(paramType)
-            }
-
-        val result =
-            ForDirectiveUtil.getFieldAccessLastPropertyClassType(
-                blockElement,
-                project,
-                topElementParentClass,
-                shortName = this.shortName,
-                isBatchAnnotation = isBatchAnnotation,
-            )
-
-        result?.highlightElement(holder)
-    }
-
-    private fun errorHighlight(
-        topElement: SqlElIdExpr,
-        daoMethod: PsiMethod,
-        holder: ProblemsHolder,
-    ) {
-        ValidationDaoParamResult(
-            topElement,
-            daoMethod.name,
-            this.shortName,
-        ).highlightElement(holder)
-    }
-
-    /**
-     * Check for existence of static field
-     */
-    private fun checkStaticFieldAndMethodAccess(
-        staticAccuser: SqlElStaticFieldAccessExpr,
-        holder: ProblemsHolder,
-    ) {
-        val blockElements = staticAccuser.accessElements
-        val psiStaticClass = PsiStaticElement(staticAccuser.elClass.elIdExprList, staticAccuser.containingFile)
-        val referenceClass = psiStaticClass.getRefClazz()
-        if (referenceClass == null) {
-            ValidationClassPathResult(
-                staticAccuser.elClass,
-                this.shortName,
-            ).highlightElement(holder)
-            return
-        }
-
-        val topParentClass = ForDirectiveUtil.getStaticFieldAccessTopElementClassType(staticAccuser, referenceClass)
-        if (topParentClass == null) {
-            blockElements.firstOrNull()?.let {
-                ValidationPropertyResult(
-                    it,
-                    PsiParentClass(referenceClass.psiClassType),
-                    this.shortName,
-                ).highlightElement(holder)
-            }
-            return
-        }
-        val result =
-            topParentClass.let {
-                ForDirectiveUtil.getFieldAccessLastPropertyClassType(
-                    blockElements,
-                    staticAccuser.project,
-                    it,
-                    shortName = this.shortName,
-                )
-            }
-        result?.highlightElement(holder)
+        val handler = InspectionPrimaryVisitorProcessor(this.shortName, element)
+        handler.check(holder, visitFile)
     }
 }
